+++ app/service/page.tsx
@@ -1,0 +1,5 @@
+import ServiceSheets from "../components/ServiceSheets";
+
+export default function ServicePage() {
+  return <ServiceSheets />;
+}